adding in the awesome date picker component, documentation not needed, 508 compliant and cross browser compatible
